+++ app/controllers/zaikio/oauth_client/subscriptions_controller.rb
@@ -1,23 +1,27 @@
 module Zaikio
   module OAuthClient
     class SubscriptionsController < ConnectionsController
-      def new
-        opts = params.permit(:client_name, :state, :plan, :organization_id)
+      def new  # rubocop:disable Metrics/MethodLength
+        opts = params.permit(:client_name, :state, :plan, :organization_id, :app_name, :redirect_uri)
         opts[:state] ||= session[:state] = SecureRandom.urlsafe_base64(32)
 
         plan            = opts.delete(:plan)
         organization_id = opts.delete(:organization_id)
+        app_name        = opts.delete(:app_name)
+        redirect_uri    = opts.delete(:redirect_uri)
 
-        subscription_scope = if organization_id.present?
-                               "Org/#{organization_id}.subscription_create"
+        scope = "Org.subscription_create"
+        scope_with_org_id = "Org/#{organization_id}.subscription_create"
+        subscription_scope = if app_name.present?
+                               organization_id.present? ? "#{scope_with_org_id}_#{app_name}" : "#{scope}_#{app_name}"
                              else
-                               "Org.subscription_create"
+                               organization_id.present? ? scope_with_org_id : scope
                              end
 
         subscription_scope << ".#{plan}" if plan.present?
 
         redirect_to oauth_client.auth_code.authorize_url(
-          redirect_uri: approve_url(opts.delete(:client_name)),
+          redirect_uri: redirect_uri || approve_url(opts.delete(:client_name)),
           scope: subscription_scope,
           **opts
         ), allow_other_host: true